+++ packages/react-wrap-with/__tests__/__types__/emptyProps.tsx
@@ -4,6 +4,6 @@ import wrapWith from '../../src/wrapWith';
 
 import type { PropsWithChildren } from 'react';
 
-const Header = ({ children }: PropsWithChildren<{}>) => <h1>{children}</h1>;
+const Header = ({ children }: PropsWithChildren<Record<any, never>>) => <h1>{children}</h1>;
 
 wrapWith(Header);